+++ client/src/app/dashboard/reports/reports.component.ts
@@ -1,3 +1,4 @@
+import { AccessGuardService } from 'src/app/Services/access-guard/access-guard.service';
 import { ReportService } from './../../Services/report/report.service';
 import { Component, OnInit } from '@angular/core';
 import { Report } from 'src/app/Objects/Report';
@@ -24,7 +25,7 @@ export class ReportsComponent implements OnInit {
   isDeletting:boolean = false;
   isEditing:boolean = false;
 
-  constructor(private reportService:ReportService) { }
+  constructor(private reportService:ReportService,private accessGuardService:AccessGuardService) { }
 
   ngOnInit(): void {
     this.reportService.get().subscribe((data: any) => {
@@ -35,18 +36,17 @@ export class ReportsComponent implements OnInit {
 
   setTableData(){
     this.reportsTable = this.reports.map(report => {
-      const status = report.status == true ? 'Active' : 'Inactive';
       return [
         {key:'image',val:report.problem?.name,path:report.problem?.icon}, // problem
         {key:'problem',val:report.problem?.name}, // problem
         {key:'map',val:[report.lat,report.lan]},  // localisation
-        {key:'status',val:status}, // status
+        {key:'boolean',val:report.status,labels:["Solved",'Pending']}, // status
         {key:'note',val:report.note?.substring(0,70)+"..."},  // note
         {key:'created_at',val:moment(report.createdAt).format('YYYY-MM-DD hh:mm:ss')}, // created_at
         {key:'actions',actions:[  // actions
           {action:"view",id:report.id},
           {action:"edit",id:report.id},
-          {action:"delete",id:report.id}
+          (this.accessGuardService.isAdmin()) ? {action:"delete",id:report.id} : null
         ]}
       ]
     })
@@ -63,6 +63,7 @@ export class ReportsComponent implements OnInit {
   }
 
   callDelete(id:any){
+    if(this.accessGuardService.isAdmin()) return;
     this.activeReport.id = id;
     this.isDeletting = true;
   }
@@ -74,6 +75,7 @@ export class ReportsComponent implements OnInit {
   }
 
   delete(){
+    if(this.accessGuardService.isAdmin()) return;
     this.reportService.delete(this.activeReport.id as number).subscribe((data: any) => {
     this.reports = this.reports.filter(report => report.id != this.activeReport.id);
     this.setTableData();
